refactored naming + hierarchy
