+++ src/StaticMock/Hooks/Entities/MethodMemoryInfoX32.cs
@@ -1,0 +1,8 @@
+﻿namespace StaticMock.Hooks.Entities;
+
+internal struct MethodMemoryInfoX32
+{
+    public byte Byte1 { get; set; }
+    public uint MethodMemoryValue { get; set; }
+    public byte Byte1AfterMethod { get; set; }
+}
