+++ src/StaticMock/Hooks/Entities/MethodMemoryInfoX64.cs
@@ -1,0 +1,11 @@
+﻿namespace StaticMock.Hooks.Entities;
+
+internal struct MethodMemoryInfoX64
+{
+    public byte Byte1 { get; set; }
+    public byte Byte2 { get; set; }
+    public ulong MethodMemoryValue { get; set; }
+    public byte Byte1AfterMethod { get; set; }
+    public byte Byte2AfterMethod { get; set; }
+    public byte Byte3AfterMethod { get; set; }
+}
