+++ src/StaticMock/Hooks/IHookManager.cs
@@ -1,0 +1,8 @@
+﻿using System.Reflection;
+
+namespace StaticMock.Hooks;
+
+internal interface IHookManager : IReturnable
+{
+    IReturnable ApplyHook(MethodBase hookMethod);
+}
