+++ src/StaticMock/Hooks/IHookManagerFactory.cs
@@ -1,0 +1,8 @@
+﻿using System.Reflection;
+
+namespace StaticMock.Hooks;
+
+internal interface IHookManagerFactory
+{
+    IHookManager CreateHookService(MethodBase originalMethod);
+}
