+++ src/StaticMock/Hooks/IReturnable.cs
@@ -1,0 +1,6 @@
+﻿namespace StaticMock.Hooks;
+
+internal interface IReturnable : IDisposable
+{
+    void Return();
+}
