+++ src/StaticMock/Hooks/Implementation/HookManagerFactory.cs
@@ -1,0 +1,11 @@
+﻿using System.Reflection;
+
+namespace StaticMock.Hooks.Implementation;
+
+internal class HookManagerFactory : IHookManagerFactory
+{
+    public IHookManager CreateHookService(MethodBase originalMethod) =>
+        IntPtr.Size == sizeof(int)
+            ? (IHookManager) new HookManagerX32(originalMethod)
+            : new HookManagerX64(originalMethod);
+}
