+++ src/StaticMock/Mock.cs
@@ -1,40 +1,40 @@
 ﻿using StaticMock.Helpers;
-using StaticMock.Services.Hook.Implementation;
-using StaticMock.Services.Mock;
-using StaticMock.Services.Mock.Implementation;
 using System.Linq.Expressions;
 using System.Reflection;
 using StaticMock.Entities;
+using StaticMock.Hooks.Implementation;
+using StaticMock.Mocks;
+using StaticMock.Mocks.Implementation;
 
 namespace StaticMock;
 
 public static class Mock
 {
-    public static IFuncMockService Setup(Type type, string methodName, Action action) => SetupMockHelper.SetupInternal(type, methodName, action);
+    public static IFuncMock Setup(Type type, string methodName, Action action) => SetupMockHelper.SetupInternal(type, methodName, action);
 
-    public static IFuncMockService Setup(Type type, string methodName, BindingFlags bindingFlags, Action action) =>
+    public static IFuncMock Setup(Type type, string methodName, BindingFlags bindingFlags, Action action) =>
         SetupMockHelper.SetupInternal(type, methodName, action, new SetupProperties
         {
             BindingFlags = bindingFlags
         });
 
-    public static IFuncMockService Setup(Type type, string methodName, SetupProperties setupProperties, Action action) =>
+    public static IFuncMock Setup(Type type, string methodName, SetupProperties setupProperties, Action action) =>
         SetupMockHelper.SetupInternal(type, methodName, action, setupProperties);
 
-    public static IFuncMockService SetupProperty(Type type, string propertyName, Action action) => SetupMockHelper.SetupPropertyInternal(type, propertyName, action);
+    public static IFuncMock SetupProperty(Type type, string propertyName, Action action) => SetupMockHelper.SetupPropertyInternal(type, propertyName, action);
 
-    public static IFuncMockService SetupProperty(Type type, string propertyName, BindingFlags bindingFlags, Action action) =>
+    public static IFuncMock SetupProperty(Type type, string propertyName, BindingFlags bindingFlags, Action action) =>
         SetupMockHelper.SetupPropertyInternal(type, propertyName, action, bindingFlags);
 
-    public static IVoidMockService SetupVoid(Type type, string methodName, Action action) => SetupMockHelper.SetupVoidInternal(type, methodName, action);
+    public static IVoidMock SetupVoid(Type type, string methodName, Action action) => SetupMockHelper.SetupVoidInternal(type, methodName, action);
 
-    public static IVoidMockService SetupVoid(Type type, string methodName, BindingFlags bindingFlags, Action action) =>
+    public static IVoidMock SetupVoid(Type type, string methodName, BindingFlags bindingFlags, Action action) =>
         SetupMockHelper.SetupVoidInternal(type, methodName, action, new SetupProperties
         {
             BindingFlags = bindingFlags
         });
 
-    public static IVoidMockService SetupVoid(Type type, string methodName, SetupProperties setupProperties, Action action) =>
+    public static IVoidMock SetupVoid(Type type, string methodName, SetupProperties setupProperties, Action action) =>
         SetupMockHelper.SetupVoidInternal(type, methodName, action, setupProperties);
 
     public static void SetupDefault(Type type, string methodName, Action action) => SetupMockHelper.SetupDefaultInternal(type, methodName, action);
@@ -48,7 +48,7 @@ public static class Mock
     public static void SetupDefault(Type type, string methodName, SetupProperties setupProperties, Action action) =>
         SetupMockHelper.SetupDefaultInternal(type, methodName, action, setupProperties);
 
-    public static IFuncMockService<TReturnValue> Setup<TReturnValue>(Expression<Func<TReturnValue>> methodGetExpression, Action action)
+    public static IFuncMock<TReturnValue> Setup<TReturnValue>(Expression<Func<TReturnValue>> methodGetExpression, Action action)
     {
         if (action == null)
         {
@@ -56,10 +56,10 @@ public static class Mock
         }
 
         var originalMethodInfo = SetupMockHelper.ValidateAndGetOriginalMethodInfo(methodGetExpression);
-        return new FuncMockService<TReturnValue>(new HookServiceFactory(), new HookBuilder(), originalMethodInfo, action);
+        return new FuncMock<TReturnValue>(new HookManagerFactory(), originalMethodInfo, action);
     }
 
-    public static IAsyncFuncMockService<TReturnValue> Setup<TReturnValue>(Expression<Func<Task<TReturnValue>>> methodGetExpression, Action action)
+    public static IAsyncFuncMock<TReturnValue> Setup<TReturnValue>(Expression<Func<Task<TReturnValue>>> methodGetExpression, Action action)
     {
         if (action == null)
         {
@@ -67,10 +67,10 @@ public static class Mock
         }
 
         var originalMethodInfo = SetupMockHelper.ValidateAndGetOriginalMethodInfo(methodGetExpression);
-        return new AsyncFuncMockService<TReturnValue>(new HookServiceFactory(), new HookBuilder(), originalMethodInfo, action);
+        return new AsyncFuncMock<TReturnValue>(new HookManagerFactory(), originalMethodInfo, action);
     }
 
-    public static IVoidMockService Setup(Expression<Action> methodGetExpression, Action action)
+    public static IVoidMock Setup(Expression<Action> methodGetExpression, Action action)
     {
         if (methodGetExpression == null)
         {
@@ -87,7 +87,7 @@ public static class Mock
             throw new Exception("Get expression not contains method to setup");
         }
 
-        return new VoidMockService(new HookServiceFactory(), new HookBuilder(), methodExpression.Method, action);
+        return new VoidMock(new HookManagerFactory(), methodExpression.Method, action);
     }
 
     public static void SetupDefault(Expression<Action> methodGetExpression, Action action)
