+++ src/StaticMock/Mocks/IVoidMock.cs
@@ -1,0 +1,6 @@
+﻿namespace StaticMock.Mocks;
+
+public interface IVoidMock : IMock
+{
+    void Callback(Action callback);
+}
