--- src/StaticMock/Services/Common/IReturnable.cs
@@ -1,6 +1,0 @@
-﻿namespace StaticMock.Services.Common;
-
-internal interface IReturnable : IDisposable
-{
-    void Return();
-}
