--- src/StaticMock/Services/Hook/Entities/MethodMemoryInfoX32.cs
@@ -1,8 +1,0 @@
-﻿namespace StaticMock.Services.Hook.Entities;
-
-internal struct MethodMemoryInfoX32
-{
-    public byte Byte1 { get; set; }
-    public uint MethodMemoryValue { get; set; }
-    public byte Byte1AfterMethod { get; set; }
-}
