--- src/StaticMock/Services/Hook/Entities/MethodMemoryInfoX64.cs
@@ -1,11 +1,0 @@
-﻿namespace StaticMock.Services.Hook.Entities;
-
-internal struct MethodMemoryInfoX64
-{
-    public byte Byte1 { get; set; }
-    public byte Byte2 { get; set; }
-    public ulong MethodMemoryValue { get; set; }
-    public byte Byte1AfterMethod { get; set; }
-    public byte Byte2AfterMethod { get; set; }
-    public byte Byte3AfterMethod { get; set; }
-}
