--- src/StaticMock/Services/Hook/IHookBuilder.cs
@@ -1,9 +1,0 @@
-﻿using System.Reflection;
-
-namespace StaticMock.Services.Hook;
-
-internal interface IHookBuilder
-{
-    MethodInfo CreateReturnHook<TReturn>(TReturn value);
-    MethodInfo CreateThrowsHook<TException>(TException exception) where TException : Exception, new();
-}
