--- src/StaticMock/Services/Hook/IHookService.cs
@@ -1,9 +1,0 @@
-﻿using System.Reflection;
-using StaticMock.Services.Common;
-
-namespace StaticMock.Services.Hook;
-
-internal interface IHookService : IReturnable
-{
-    IReturnable Hook(MethodBase hookMethod);
-}
