--- src/StaticMock/Services/Hook/IHookServiceFactory.cs
@@ -1,8 +1,0 @@
-﻿using System.Reflection;
-
-namespace StaticMock.Services.Hook;
-
-internal interface IHookServiceFactory
-{
-    IHookService CreateHookService(MethodBase originalMethod);
-}
