--- src/StaticMock/Services/Mock/IAsyncFuncMockService.cs
@@ -1,7 +1,0 @@
-﻿namespace StaticMock.Services.Mock;
-
-public interface IAsyncFuncMockService<TReturnValue> : IFuncMockService<Task<TReturnValue>>
-{
-    void ReturnsAsync(TReturnValue value);
-    void CallbackAsync(Func<TReturnValue> callback);
-}
