--- src/StaticMock/Services/Mock/IVoidMockService.cs
@@ -1,6 +1,0 @@
-﻿namespace StaticMock.Services.Mock;
-
-public interface IVoidMockService : IMockService
-{
-    void Callback(Action callback);
-}
