--- src/StaticMock/Services/Returns/IReturnsMockService.cs
@@ -1,9 +1,0 @@
-﻿using StaticMock.Services.Common;
-
-namespace StaticMock.Services.Returns;
-
-internal interface IReturnsMockService<in TValue>
-{
-    IReturnable Returns(TValue value);
-    IReturnable ReturnsAsync(TValue value);
-}
