--- src/StaticMock/Services/Throws/IThrowsService.cs
@@ -1,9 +1,0 @@
-﻿using StaticMock.Services.Common;
-
-namespace StaticMock.Services.Throws;
-
-internal interface IThrowsService
-{
-    IReturnable Throws(Type exceptionType, object[] constructorArgs);
-    IReturnable Throws<TException>() where TException : Exception, new();
-}
